Extract the PostgreSQL plan-tree walkers so a second fixture can use them
